Re-org jobs encoding/decoding
